+++ tests/test_coordinates.py
@@ -3,15 +3,13 @@
 import unittest
 
 from coaster.db import db
-from coaster.sqlalchemy import BaseMixin, LongitudeLatitudeColumns
+from coaster.sqlalchemy import BaseMixin, CoordinatesMixin
 
 from test_models import app1, app2
 
 
-class CoordinatesData(BaseMixin, db.Model):
+class CoordinatesData(BaseMixin, CoordinatesMixin, db.Model):
     __tablename__ = 'coordinates_data'
-    coordinates = LongitudeLatitudeColumns()
-    coord2 = LongitudeLatitudeColumns('coord2')
 
 
 # -- Tests --------------------------------------------------------------------
@@ -33,15 +31,12 @@ class TestCoordinatesColumn(unittest.TestCase):
 
     def test_columns_created(self):
         table = CoordinatesData.__table__
-        self.assertTrue(isinstance(table.c.longitude.type, db.Numeric))
         self.assertTrue(isinstance(table.c.latitude.type, db.Numeric))
-        self.assertTrue(isinstance(table.c.coord2_longitude.type, db.Numeric))
-        self.assertTrue(isinstance(table.c.coord2_longitude.type, db.Numeric))
+        self.assertTrue(isinstance(table.c.longitude.type, db.Numeric))
 
-    def test_column_when_null(self):
+    def test_columns_when_null(self):
         data = CoordinatesData()
         self.assertEqual(data.coordinates, (None, None))
-        self.assertEqual(data.coord2, (None, None))
 
     def test_column_set_value(self):
         data = CoordinatesData()
@@ -52,7 +47,6 @@ class TestCoordinatesColumn(unittest.TestCase):
 
         readdata = CoordinatesData.query.first()
         self.assertEqual(readdata.coordinates, (12, 73))
-        self.assertEqual(readdata.coord2, (None, None))
 
 
 class TestCoordinatesColumn2(TestCoordinatesColumn):